requirements.txt
# Core dependencies
polars>=0.20.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Configuration and experiment management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Experiment tracking (optional)
wandb>=0.15.0

# Task automation
invoke>=2.0.0

# Progress bars and visualization
tqdm>=4.65.0
matplotlib>=3.7.0

# NLP and semantic features
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# Graph features (optional, for graph-based features)
networkx>=3.1
node2vec>=0.4.6
gensim>=4.3.0

# Utilities
pyyaml>=6.0
